- Moved find_path to Utilities/misc, until we are sure that we want it as part of the GenericNetwork code - Moved units tests too
